updated utils to async await

+++ frontend/util/server_api_util.js
@@ -1,11 +1,13 @@
 export const ServerApi = {
-    list: (userId) => (
-        $.ajax({
+    list: async (userId) => {
+        
+        const res = await $.ajax({
             method: 'GET',
             url: 'api/servers',
             data: { user_id: userId }
-        }) 
-    ),
+        })
+        return res;
+    },
 
     grab: async (id) => {
         if (!id) return {};
@@ -17,26 +19,32 @@ export const ServerApi = {
         return res;
     },
     
-    create: (server) => (
-        $.ajax({
+    create: async (server) => {
+        if (!server) return {};
+        const res = await $.ajax({
             method: 'POST',
             url: 'api/servers',
             data: { server }
         })
-    ),
+        return res
+    },
 
-    update: (server) => (
-        $.ajax({
+    update: async (server) => {
+        if (!server) return {}
+        const res = await $.ajax({
             method: 'PATCH',
             url: `api/servers/${server.id}`,
             data: { server }
         })
-    ),
+        return res
+    },
     
-    delete: (id) => (
-        $.ajax({
+    delete: async (id) => {
+        if (!id) return {}
+        const res = await $.ajax({
             method: 'DELETE',
             url: `api/servers/${id}`,
         })
-    )
+        return res;
+    }
 }
